+++ libraries/template-haskell/changelog.md
@@ -14,6 +14,11 @@
     with Template Haskell. This is not a part of the public API, and as
     such, there are no API guarantees for this module from version to version.
 
+  * `MonadIO` is now a superclass of `Quasi`, `qRunIO` has a default
+    implementation `qRunIO = liftIO`
+
+  * Add `MonadIO Q` instance
+
 ## 2.12.0.0 *TBA*
 
   * Bundled with GHC *TBA*